vw: ci/cd pipeline changes
* vw: small fixes

* vw: many lint fixes

* vw: remove local claude settings from git

* vw: fix gitignore

* vw: cleanup repo and fix analyze pipeline

* vw: fix analyze pipeline

* vw: remove clang tidy from analysis

* vw: pipeline job rename

+++ apps/sculptor/src/app/app_state.h
@@ -15,7 +15,7 @@ namespace vw::sculptor {
 
 using keyframe_value = std::variant<gfx::keyframe_vec3f, gfx::keyframe_quat>;
 
-enum class tools : int {
+enum class tools : uint8 {
     invalid,
     add_voxel,
     remove_voxel,
